De-rebenchify haste.
We've noticed that rebench doesn't spawn a fresh process for each
invocation of a benchmark. As a result, if your benchmark has a lot of
performance variation at the process-level, then you "pin" the
performance that you happen to get for the first invocation across all
invocations.

This change replaces rebench with a simple runner that does spawn new
processes.

Rather than inheriting rebench's YAML config file, I've made our own
(much simplified) TOML config file that is the bare minimum to yk's
requirements.

Similarly, the results are now stored in TOML too for ease of
reading/writing in Rust.

Also for simplicity, haste now measures/summarises the time taken for a
process execution (what haste calls an invocation) instead of an
in-process iteration (what haste calls a iteration). This is good enough
for us and simplifies everything.

Further, added a small example which doubles up as a minimal integration
test.