Output on a single line.
This changes output so that each benchmark only appears on one line,
using `\r` to change the output.

For example when running a benchmark, a line in output will evolve as
follows:

```
>>> haste:   0% deltablue/yklua/12000 ...
>>> haste:   4% deltablue/yklua/12000 11109ms ...
>>> haste:   4% deltablue/yklua/12000 11109ms 10376ms ...
>>> haste:   4% deltablue/yklua/12000 11109ms 10376ms 10540ms
```

This makes the output a lot less chatty, and easier to compare to
previous runs on the fly.